# mobile/contracts/mobile-api.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MobileCredentialLoginRequest": {
        "properties": {
          "device": {
            "$ref": "#/components/schemas/MobileDeviceInput"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "device"
        ],
        "title": "MobileCredentialLoginRequest",
        "type": "object"
      },
      "MobileDeviceInput": {
        "properties": {
          "app_version": {
            "maxLength": 40,
            "minLength": 1,
            "title": "App Version",
            "type": "string"
          },
          "device_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "installation_id": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Installation Id",
            "type": "string"
          },
          "platform": {
            "enum": [
              "android",
              "ios"
            ],
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "installation_id",
          "platform",
          "app_version"
        ],
        "title": "MobileDeviceInput",
        "type": "object"
      },
...
      "MobileLogoutRequest": {
        "properties": {
          "refresh_token": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Token"
          }
        },
        "title": "MobileLogoutRequest",
        "type": "object"
      },
...
      "MobilePrincipalResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "agency_id": {
            "format": "uuid",
            "title": "Agency Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "force_password_change": {
            "default": false,
            "title": "Force Password Change",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "passenger_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passenger Id"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "principal_type": {
            "enum": [
              "passenger",
              "client_manager",
              "coordinator"
            ],
            "title": "Principal Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "principal_type",
          "agency_id",
          "display_name"
        ],
        "title": "MobilePrincipalResponse",
        "type": "object"
      },
...
      "MobileRefreshRequest": {
        "properties": {
          "installation_id": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Installation Id",
            "type": "string"
          },
          "refresh_token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token",
          "installation_id"
        ],
        "title": "MobileRefreshRequest",
        "type": "object"
      },
...
      "MobileTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "access_token_expires_at": {
            "format": "date-time",
            "title": "Access Token Expires At",
            "type": "string"
          },
          "offline_authorization_lease": {
            "maxLength": 4096,
            "minLength": 256,
            "pattern": "^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$",
            "title": "Offline Authorization Lease",
            "type": "string"
          },
          "principal": {
            "$ref": "#/components/schemas/MobilePrincipalResponse"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "refresh_token_expires_at": {
            "format": "date-time",
            "title": "Refresh Token Expires At",
            "type": "string"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "access_token_expires_at",
          "refresh_token_expires_at",
          "session_id",
          "offline_authorization_lease",
          "principal"
        ],
        "title": "MobileTokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Global Connects Dashboard",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mobile/auth/login": {
      "post": {
        "operationId": "mobile_credential_login_api_v1_mobile_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileCredentialLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mobile Credential Login",
        "tags": [
          "GC Mobile Authentication"
        ]
      }
    },
    "/api/v1/mobile/auth/logout": {
      "post": {
        "operationId": "logout_mobile_session_api_v1_mobile_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MobileLogoutRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout Mobile Session",
        "tags": [
          "GC Mobile Authentication"
        ]
      }
    },
    "/api/v1/mobile/auth/logout-all": {
      "post": {
        "operationId": "logout_all_mobile_sessions_api_v1_mobile_auth_logout_all_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout All Mobile Sessions",
        "tags": [
          "GC Mobile Authentication"
        ]
      }
    },
    "/api/v1/mobile/auth/me": {
      "get": {
        "operationId": "mobile_me_api_v1_mobile_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobilePrincipalResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mobile Me",
        "tags": [
          "GC Mobile Authentication"
        ]
      }
    },
...
    "/api/v1/mobile/auth/refresh": {
      "post": {
        "operationId": "refresh_mobile_session_api_v1_mobile_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MobileRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Mobile Session",
        "tags": [
          "GC Mobile Authentication"
        ]
      }
    },
...
  }
}